requirements.txt
mysql-connector-python==8.0.31
googlemaps==
numpy==1.23.4
pandas==1.1.4
PyMySQL==0.10.1
python-dotenv==0.21.0
